World: Lock save file while accessing it
Add a mutex to lock the save file when accessing it to prevent
corruption and race conditions.
