+++ test/thread_tests.cpp
@@ -12,6 +12,7 @@
 #include <gtest/gtest.h>
 
 #include "thread/TestKanoWorldRefresh.h"
+#include "thread/TestKanoWorldSave.h"
 
 
 int main(int argc, char *argv[]) {
